requirements.txt
python-dotenv==1.0.1
discord==2.3.2
discord.py==2.3.2
yt-dlp==2024.3.10
PyNaCl==1.5.0
